+++ src/chain_sync/chain_muxer.rs
@@ -26,10 +26,10 @@ use futures::{
     try_join, StreamExt,
 };
 use fvm_ipld_blockstore::Blockstore;
-use log::{debug, error, info, trace, warn};
 use parking_lot::RwLock;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
+use tracing::{debug, error, info, trace, warn};
 
 use crate::chain_sync::{
     bad_block_cache::BadBlockCache,
@@ -307,7 +307,7 @@ where
 
         let epoch = block.header.epoch();
 
-        log::debug!(
+        debug!(
             "Getting messages of gossipblock, epoch: {epoch}, block: {}",
             block.header.cid()
         );
@@ -315,14 +315,14 @@ where
         let bls_messages: Vec<_> = block
             .bls_messages
             .into_iter()
-            .map(|m| network.bitswap_get::<Message>(epoch, m))
+            .map(|m| network.bitswap_get::<Message>(m))
             .collect();
 
         // Get secp_messages in the store or over Bitswap
         let secp_messages: Vec<_> = block
             .secpk_messages
             .into_iter()
-            .map(|m| network.bitswap_get::<SignedMessage>(epoch, m))
+            .map(|m| network.bitswap_get::<SignedMessage>(m))
             .collect();
 
         let (bls_messages, secp_messages) =
@@ -336,7 +336,7 @@ where
             bls_messages,
             secp_messages,
         };
-        Ok(FullTipset::new(vec![block]).unwrap())
+        Ok(FullTipset::from(block))
     }
 
     fn handle_pubsub_message(mem_pool: Arc<MessagePool<M>>, message: SignedMessage) {