+++ src/daemon/mod.rs
@@ -41,7 +41,6 @@ use bundle::load_bundles;
 use dialoguer::{console::Term, theme::ColorfulTheme};
 use futures::{select, Future, FutureExt};
 use lazy_static::lazy_static;
-use log::{debug, info, warn};
 use raw_sync::events::{Event, EventInit as _, EventState};
 use shared_memory::ShmemConf;
 use std::{
@@ -61,6 +60,7 @@ use tokio::{
     sync::{mpsc, RwLock},
     task::JoinSet,
 };
+use tracing::{debug, info, warn};
 
 lazy_static! {
     static ref IPC_PATH: TempPath = Builder::new()
@@ -366,7 +366,13 @@ pub(super) async fn start(
         services.spawn(async move {
             info!("JSON-RPC endpoint started at {}", config.client.rpc_address);
             // XXX: The JSON error message are a nightmare to print.
-            start_rpc::<_, _, cns::FullConsensus>(
+            let beacon = Arc::new(
+                rpc_state_manager
+                    .chain_config()
+                    .get_beacon_schedule(chain_store.genesis()?.timestamp())
+                    .into_dyn(),
+            );
+            start_rpc(
                 Arc::new(RPCState {
                     state_manager: Arc::clone(&rpc_state_manager),
                     keystore: keystore_rpc,
@@ -377,7 +383,7 @@ pub(super) async fn start(
                     network_name,
                     start_time,
                     // TODO: the RPCState can fetch this itself from the StateManager
-                    beacon: rpc_state_manager.beacon_schedule(),
+                    beacon,
                     chain_store: rpc_chain_store,
                     new_mined_block_tx: tipset_sink,
                     gc_event_tx,
@@ -413,6 +419,8 @@ pub(super) async fn start(
             &state_manager,
             &path.display().to_string(),
             config.client.skip_load,
+            config.client.chunk_size,
+            config.client.buffer_size,
         )
         .await
         .context("Failed miserably while importing chain from snapshot")?;
@@ -430,8 +438,10 @@ pub(super) async fn start(
         assert!(current_height.is_positive());
         match validate_from.is_negative() {
             // allow --height=-1000 to scroll back from the current head
-            true => state_manager.validate((current_height + validate_from)..=current_height)?,
-            false => state_manager.validate(validate_from..=current_height)?,
+            true => {
+                state_manager.validate_range((current_height + validate_from)..=current_height)?
+            }
+            false => state_manager.validate_range(validate_from..=current_height)?,
         }
     }
 
@@ -504,14 +514,12 @@ async fn fetch_snapshot_if_required(
                 crate::cli_shared::snapshot::peek(vendor, &config.chain.network)
                     .await
                     .context("couldn't get snapshot size")?;
-            let num_bytes = byte_unit::Byte::from(num_bytes)
-                .get_appropriate_unit(true)
-                .format(2);
             // dialoguer will double-print long lines, so manually print the first clause ourselves,
             // then let `Confirm` handle the second.
             println!("Forest requires a snapshot to sync with the network, but automatic fetching is disabled.");
             let message = format!(
-                "Fetch a {num_bytes} snapshot to the current directory? (denying will exit the program). "
+                "Fetch a {} snapshot to the current directory? (denying will exit the program). ",
+                indicatif::HumanBytes(num_bytes)
             );
             let have_permission = asyncify(|| {
                 dialoguer::Confirm::with_theme(&ColorfulTheme::default())
@@ -715,6 +723,7 @@ fn create_password(prompt: &str) -> std::io::Result<String> {
 #[cfg(test)]
 mod test {
     use crate::blocks::BlockHeader;
+    use crate::cli_shared::cli::{BufferSize, ChunkSize};
     use crate::db::MemoryDB;
     use crate::networks::ChainConfig;
     use crate::shim::address::Address;
@@ -774,7 +783,14 @@ mod test {
             chain_config,
             Arc::new(crate::interpreter::RewardActorMessageCalc),
         )?);
-        import_chain::<_>(&sm, file_path, false).await?;
+        import_chain::<_>(
+            &sm,
+            file_path,
+            false,
+            ChunkSize::default(),
+            BufferSize::default(),
+        )
+        .await?;
         Ok(())
     }
 
@@ -799,9 +815,15 @@ mod test {
             chain_config,
             Arc::new(crate::interpreter::RewardActorMessageCalc),
         )?);
-        import_chain::<_>(&sm, "test-snapshots/chain4.car", false)
-            .await
-            .context("Failed to import chain")?;
+        import_chain::<_>(
+            &sm,
+            "test-snapshots/chain4.car",
+            false,
+            ChunkSize::default(),
+            BufferSize::default(),
+        )
+        .await
+        .context("Failed to import chain")?;
 
         Ok(())
     }