+++ src/interpreter/mod.rs
@@ -1,12 +1,9 @@
 // Copyright 2019-2023 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
+mod errors;
 mod fvm;
 pub mod fvm3;
-#[cfg(feature = "instrumented_kernel")]
-mod instrumented_kernel;
-#[cfg(feature = "instrumented_kernel")]
-mod metrics;
 mod vm;
 
 use crate::shim::{