+++ src/rpc/mpool_api.rs
@@ -4,7 +4,6 @@
 
 use std::convert::TryFrom;
 
-use crate::beacon::Beacon;
 use crate::blocks::TipsetKeys;
 use crate::json::{
     cid::{vec::CidJsonVec, CidJson},
@@ -16,19 +15,17 @@ use crate::rpc_api::{data_types::RPCState, mpool_api::*};
 use crate::shim::address::Protocol;
 use ahash::{HashSet, HashSetExt};
 use fvm_ipld_blockstore::Blockstore;
-use fvm_ipld_encoding::Cbor;
 use jsonrpc_v2::{Data, Error as JsonRpcError, Params};
 
 use super::gas_api::estimate_message_gas;
 
 /// Return `Vec` of pending messages in `mpool`
-pub(in crate::rpc) async fn mpool_pending<DB, B>(
-    data: Data<RPCState<DB, B>>,
+pub(in crate::rpc) async fn mpool_pending<DB>(
+    data: Data<RPCState<DB>>,
     Params(params): Params<MpoolPendingParams>,
 ) -> Result<MpoolPendingResult, JsonRpcError>
 where
     DB: Blockstore + Clone + Send + Sync + 'static,
-    B: Beacon,
 {
     let (CidJsonVec(cid_vec),) = params;
     let tsk = TipsetKeys::new(cid_vec);
@@ -82,13 +79,12 @@ where
 }
 
 /// Add `SignedMessage` to `mpool`, return message CID
-pub(in crate::rpc) async fn mpool_push<DB, B>(
-    data: Data<RPCState<DB, B>>,
+pub(in crate::rpc) async fn mpool_push<DB>(
+    data: Data<RPCState<DB>>,
     Params(params): Params<MpoolPushParams>,
 ) -> Result<MpoolPushResult, JsonRpcError>
 where
     DB: Blockstore + Clone + Send + Sync + 'static,
-    B: Beacon,
 {
     let (SignedMessageJson(smsg),) = params;
 
@@ -98,13 +94,12 @@ where
 }
 
 /// Sign given `UnsignedMessage` and add it to `mpool`, return `SignedMessage`
-pub(in crate::rpc) async fn mpool_push_message<DB, B>(
-    data: Data<RPCState<DB, B>>,
+pub(in crate::rpc) async fn mpool_push_message<DB>(
+    data: Data<RPCState<DB>>,
     Params(params): Params<MpoolPushMessageParams>,
 ) -> Result<MpoolPushMessageResult, JsonRpcError>
 where
     DB: Blockstore + Clone + Send + Sync + 'static,
-    B: Beacon,
 {
     let (MessageJson(umsg), spec) = params;
 
@@ -122,7 +117,7 @@ where
             "Expected nonce for MpoolPushMessage is 0, and will be calculated for you.".into(),
         );
     }
-    let mut umsg = estimate_message_gas::<DB, B>(&data, umsg, spec, Default::default()).await?;
+    let mut umsg = estimate_message_gas::<DB>(&data, umsg, spec, Default::default()).await?;
     if umsg.gas_premium > umsg.gas_fee_cap {
         return Err("After estimation, gas premium is greater than gas fee cap".into());
     }