+++ src/rpc/sync_api.rs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 #![allow(clippy::unused_async)]
 
-use crate::beacon::Beacon;
 use crate::chain_sync::SyncState;
 use crate::json::cid::CidJson;
 use crate::rpc_api::{
@@ -14,26 +13,24 @@ use jsonrpc_v2::{Data, Error as JsonRpcError, Params};
 use parking_lot::RwLock;
 
 /// Checks if a given block is marked as bad.
-pub(in crate::rpc) async fn sync_check_bad<DB, B>(
-    data: Data<RPCState<DB, B>>,
+pub(in crate::rpc) async fn sync_check_bad<DB>(
+    data: Data<RPCState<DB>>,
     Params(params): Params<SyncCheckBadParams>,
 ) -> Result<SyncCheckBadResult, JsonRpcError>
 where
     DB: Blockstore,
-    B: Beacon,
 {
     let (CidJson(cid),) = params;
     Ok(data.bad_blocks.peek(&cid).unwrap_or_default())
 }
 
 /// Marks a block as bad, meaning it will never be synced.
-pub(in crate::rpc) async fn sync_mark_bad<DB, B>(
-    data: Data<RPCState<DB, B>>,
+pub(in crate::rpc) async fn sync_mark_bad<DB>(
+    data: Data<RPCState<DB>>,
     Params(params): Params<SyncMarkBadParams>,
 ) -> Result<SyncMarkBadResult, JsonRpcError>
 where
     DB: Blockstore,
-    B: Beacon,
 {
     let (CidJson(cid),) = params;
     data.bad_blocks
@@ -46,12 +43,11 @@ async fn clone_state(state: &RwLock<SyncState>) -> SyncState {
 }
 
 /// Returns the current status of the `ChainSync` process.
-pub(in crate::rpc) async fn sync_state<DB, B>(
-    data: Data<RPCState<DB, B>>,
+pub(in crate::rpc) async fn sync_state<DB>(
+    data: Data<RPCState<DB>>,
 ) -> Result<SyncStateResult, JsonRpcError>
 where
     DB: Blockstore,
-    B: Beacon,
 {
     let active_syncs = vec![clone_state(data.sync_state.as_ref()).await];
     Ok(RPCSyncState { active_syncs })
@@ -72,7 +68,7 @@ mod tests {
     use crate::networks::ChainConfig;
     use crate::shim::address::Address;
     use crate::state_manager::StateManager;
-    use fvm_ipld_encoding::Cbor;
+    use fvm_ipld_encoding::from_slice;
     use serde_json::from_str;
     use tempfile::TempDir;
     use tokio::{sync::RwLock, task::JoinSet};
@@ -81,14 +77,14 @@ mod tests {
 
     const TEST_NET_NAME: &str = "test";
 
-    fn state_setup() -> (
-        Arc<RPCState<MemoryDB, MockBeacon>>,
-        flume::Receiver<NetworkMessage>,
-    ) {
-        let beacon = Arc::new(BeaconSchedule(vec![BeaconPoint {
-            height: 0,
-            beacon: Arc::new(MockBeacon::default()),
-        }]));
+    fn state_setup() -> (Arc<RPCState<MemoryDB>>, flume::Receiver<NetworkMessage>) {
+        let beacon = Arc::new(
+            BeaconSchedule(vec![BeaconPoint {
+                height: 0,
+                beacon: MockBeacon::default(),
+            }])
+            .into_dyn(),
+        );
 
         let (network_send, network_rx) = flume::bounded(5);
         let mut services = JoinSet::new();
@@ -127,7 +123,7 @@ mod tests {
         let mpool_network_send = network_send.clone();
         let pool = {
             let bz = hex::decode("904300e80781586082cb7477a801f55c1f2ea5e5d1167661feea60a39f697e1099af132682b81cc5047beacf5b6e80d5f52b9fd90323fb8510a5396416dd076c13c85619e176558582744053a3faef6764829aa02132a1571a76aabdc498a638ea0054d3bb57f41d82015860812d2396cc4592cdf7f829374b01ffd03c5469a4b0a9acc5ccc642797aa0a5498b97b28d90820fedc6f79ff0a6005f5c15dbaca3b8a45720af7ed53000555667207a0ccb50073cd24510995abd4c4e45c1e9e114905018b2da9454190499941e818201582012dd0a6a7d0e222a97926da03adb5a7768d31cc7c5c2bd6828e14a7d25fa3a608182004b76616c69642070726f6f6681d82a5827000171a0e4022030f89a8b0373ad69079dbcbc5addfe9b34dce932189786e50d3eb432ede3ba9c43000f0001d82a5827000171a0e4022052238c7d15c100c1b9ebf849541810c9e3c2d86e826512c6c416d2318fcd496dd82a5827000171a0e40220e5658b3d18cd06e1db9015b4b0ec55c123a24d5be1ea24d83938c5b8397b4f2fd82a5827000171a0e4022018d351341c302a21786b585708c9873565a0d07c42521d4aaf52da3ff6f2e461586102c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001a5f2c5439586102b5cd48724dce0fec8799d77fd6c5113276e7f470c8391faa0b5a6033a3eaf357d635705c36abe10309d73592727289680515afd9d424793ba4796b052682d21b03c5c8a37d94827fecc59cdc5750e198fdf20dee012f4d627c6665132298ab95004500053724e0").unwrap();
-            let header = BlockHeader::unmarshal_cbor(&bz).unwrap();
+            let header = from_slice::<BlockHeader>(&bz).unwrap();
             let ts = Tipset::from(header);
             let db = cs_for_test.blockstore();
             let tsk = ts.key().cids.clone();