+++ CHANGELOG.md
@@ -29,7 +29,19 @@
 
 ### Added
 
+- [#3167](https://github.com/ChainSafe/forest/pull/3167): Added a new option
+  `--validate-tipsets` for `forest-cli snapshot validate`.
+- [#3166](https://github.com/ChainSafe/forest/issues/3166): Add
+  `forest-cli archive info` command for inspecting archives.
+- [#3159](https://github.com/ChainSafe/forest/issues/3159): Add
+  `forest-cli archive export -e=X` command for exporting archives.
+- [#3148](https://github.com/ChainSafe/forest/pull/3148): add `save_to_file`
+  option to `forest-cli state fetch` command.
+
 ### Changed
+
+- [#3126](https://github.com/ChainSafe/forest/issues/3126): Bail on database
+  lookup errors instead of silently ignoring them.
 
 ### Removed